refactor(coupling): move MatchMethod and key validation into domain
Break domain→match and validate→hotkeys cycles by owning wire-format
MatchMethod and pure hotkey chord rules in sqyre-domain; match and
hotkeys re-export, validate calls domain directly.
