+++ crates/sqyre-hotkeys/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 
 [dependencies]
 parking_lot = { workspace = true }
+sqyre-domain = { path = "../sqyre-domain" }
 rdev = { workspace = true, optional = true }
 web-time = { workspace = true }
