+++ crates/sqyre-hotkeys/src/continue_wait.rs
@@ -1,6 +1,7 @@
 //! Shared continue-chord wait used by Pause and NavigateSelect.
 
 use parking_lot::{Condvar, Mutex};
+use sqyre_domain::{normalize_keys, validate_not_failsafe};
 use std::collections::HashSet;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
@@ -230,98 +231,6 @@ impl ContinueWaitBridge {
         }
         result
     }
-}
-
-fn normalize_keys(keys: &[String]) -> Vec<String> {
-    keys.iter()
-        .map(|k| normalize_key_name(k))
-        .filter(|k| !k.is_empty())
-        .collect()
-}
-
-pub fn normalize_key_name(key: &str) -> String {
-    match key.trim().to_ascii_lowercase().as_str() {
-        "escape" => "esc".into(),
-        "control" | "rcontrol" | "controlleft" | "controlright" => "ctrl".into(),
-        "return" => "enter".into(),
-        "super" | "meta" | "win" | "windows" | "meta_left" | "metaleft" => "cmd".into(),
-        "meta_right" | "metaright" => "rcmd".into(),
-        "backspace" | "back_space" => "delete".into(),
-        "shiftleft" | "shift_left" => "shift".into(),
-        "shiftright" | "shift_right" => "rshift".into(),
-        "altleft" | "alt_left" => "alt".into(),
-        "altright" | "alt_right" | "altgr" => "ralt".into(),
-        other => other.to_string(),
-    }
-}
-
-/// Emergency-stop chord keys (order-independent): Ctrl + Alt + Shift + Esc.
-///
-/// Includes Alt so the chord does not collide with Windows Task Manager
-/// (`Ctrl+Shift+Esc`), which the OS often delivers outside user-mode hooks.
-pub const FAILSAFE_KEYS: &[&str] = &["alt", "ctrl", "esc", "shift"];
-
-/// Human-readable failsafe chord for UI / logs.
-pub const FAILSAFE_LABEL: &str = "Esc+Ctrl+Alt+Shift";
-
-/// Collapse left/right modifier variants for failsafe matching.
-fn canonicalize_failsafe_key(key: &str) -> &str {
-    match key {
-        "rshift" => "shift",
-        "ralt" => "alt",
-        other => other,
-    }
-}
-
-/// Whether `keys` (already normalized) match the failsafe chord.
-///
-/// Left/right Shift and Alt are treated as equivalent.
-pub fn is_failsafe_chord(keys: &[String]) -> bool {
-    let mut sorted: Vec<&str> = keys
-        .iter()
-        .map(|k| canonicalize_failsafe_key(k.as_str()))
-        .collect();
-    sorted.sort_unstable();
-    sorted.dedup();
-    if sorted.len() != FAILSAFE_KEYS.len() {
-        return false;
-    }
-    let mut failsafe = FAILSAFE_KEYS.to_vec();
-    failsafe.sort_unstable();
-    sorted == failsafe
-}
-
-/// Whether Ctrl, Alt, and Shift are all held (left/right variants accepted).
-pub fn failsafe_modifiers_held(pressed: &std::collections::HashSet<&'static str>) -> bool {
-    let ctrl = pressed.contains("ctrl");
-    let alt = pressed.contains("alt") || pressed.contains("ralt");
-    let shift = pressed.contains("shift") || pressed.contains("rshift");
-    ctrl && alt && shift
-}
-
-/// Normalize and validate a Pause continue-key chord.
-///
-/// Returns normalized key names, or an error if empty / equals the failsafe chord.
-pub fn validate_continue_key(keys: &[String]) -> Result<Vec<String>, String> {
-    let normalized = normalize_keys(keys);
-    if normalized.is_empty() {
-        return Err("pause: continue key not set".into());
-    }
-    if is_failsafe_chord(&normalized) {
-        return Err(format!(
-            "pause: continue key cannot match the failsafe hotkey ({FAILSAFE_LABEL})"
-        ));
-    }
-    Ok(normalized)
-}
-
-fn validate_not_failsafe(keys: &[String]) -> Result<(), String> {
-    if is_failsafe_chord(keys) {
-        return Err(format!(
-            "key wait: chord cannot match the failsafe hotkey ({FAILSAFE_LABEL})"
-        ));
-    }
-    Ok(())
 }
 
 /// Map an rdev key to a Sqyre hotkey name (lowercase).
@@ -588,35 +497,4 @@ mod tests {
         handle.join().unwrap();
     }
 
-    #[test]
-    fn failsafe_chord_requires_alt() {
-        assert!(!is_failsafe_chord(&[
-            "ctrl".into(),
-            "esc".into(),
-            "shift".into(),
-        ]));
-        assert!(is_failsafe_chord(&[
-            "ctrl".into(),
-            "alt".into(),
-            "esc".into(),
-            "shift".into(),
-        ]));
-        assert!(is_failsafe_chord(&[
-            "ctrl".into(),
-            "ralt".into(),
-            "esc".into(),
-            "rshift".into(),
-        ]));
-    }
-
-    #[test]
-    fn failsafe_modifiers_accept_left_right() {
-        let mut pressed = HashSet::new();
-        pressed.insert("ctrl");
-        pressed.insert("ralt");
-        pressed.insert("rshift");
-        assert!(failsafe_modifiers_held(&pressed));
-        pressed.remove("ralt");
-        assert!(!failsafe_modifiers_held(&pressed));
-    }
 }
