+++ crates/sqyre-hotkeys/src/lib.rs
@@ -4,9 +4,10 @@ mod continue_wait;
 mod macro_hotkeys;
 mod screen_click;
 
-pub use continue_wait::{
-    failsafe_modifiers_held, is_failsafe_chord, normalize_key_name, validate_continue_key,
-    ContinueWaitBridge, FAILSAFE_KEYS, FAILSAFE_LABEL,
+pub use continue_wait::ContinueWaitBridge;
+pub use sqyre_domain::{
+    failsafe_modifiers_held, is_failsafe_chord, normalize_key_name, normalize_keys,
+    validate_continue_key, validate_not_failsafe, FAILSAFE_KEYS, FAILSAFE_LABEL,
 };
 pub use macro_hotkeys::{
     chord_all_pressed, chord_fully_released, format_hotkey, parse_hotkey, HotkeyTrigger,
