+++ crates/sqyre-match/Cargo.toml
@@ -14,7 +14,7 @@ pulp = { workspace = true }
 purecv = { workspace = true }
 rayon = { workspace = true }
 rustfft = { workspace = true }
-serde = { workspace = true }
 thiserror = { workspace = true }
+sqyre-domain = { path = "../sqyre-domain" }
 
 [dev-dependencies]
